refactor: extract create entry points from build config
